reduce jank on map pins load
Grabbing pins beforehand and moving the (somewhat expensive) clustering operation outside of the pin creation loop seems to alleviate some of the problems with view freezes when 100+ pins are loaded. Much work remains to be done in trying to reduce the cost of the pin merging procedure and proper culling/spatial partitioning/on-demand loading.